Clarification on a couple of points.
Blocked the text a little more into "the pieces" and "the interactions" as paragraph 1 and 2 respectively. 

I'm not happy with the way the line "At a predefined interval, known as the scaling interval, the App Autoscaler app ..." reads, but I can't think of anything better. The rub is that the audience for this doesn't necessarily have any control over the scaling interval (that's a more operations-level thing) but might benefit from knowing that it's on a consistent cycle. Feel free to edit if something better comes to mind.